fix: harden replication framing decoder and auth prefix upper bound
The default IReplicationBatchEncoder framing decoder validated entry-body and routing-field lengths with a narrowing int 'cursor + length' sum. A forged length near int.MaxValue overflowed to negative and slipped past the truncation guard, downgrading the fail-closed ArgumentException framing rejection into a raw slice throw. Widen both sums to long, mirroring the sibling fix in OrleansBinaryReplicationBatchEncoder.

LatticeAuthAdmin.PrefixUpperBound incremented the prefix's final code unit unconditionally, so a prefix ending in U+FFFF wrapped to U+0000 and produced an upper bound sorting below the prefix, inverting the [prefix, bound) range. Scan back to the last unit below char.MaxValue, drop trailing max units, and return null when unbounded, mirroring BackupConstants.PrefixUpperBound.

Adds regression tests for all three paths and a CHANGELOG entry.

+++ src/lattice.replication/IReplicationBatchEncoder.cs
@@ -343,7 +343,12 @@ public interface IReplicationBatchEncoder
             }
             var length = BinaryPrimitives.ReadInt32LittleEndian(span[cursor..(cursor + 4)]);
             cursor += 4;
-            if (length < 0 || cursor + length > payload.Length)
+            // Widen to long before summing: length is an attacker-controllable
+            // wire field (up to int.MaxValue), so cursor + length can overflow to
+            // a negative value and slip past this guard, downgrading the precise
+            // framing rejection into a raw slice exception. Mirrors the fix in
+            // OrleansBinaryReplicationBatchEncoder.
+            if (length < 0 || (long)cursor + length > payload.Length)
             {
                 throw new ArgumentException(
                     $"Framing payload is truncated at the body for entry {i} of "
@@ -375,7 +380,12 @@ public interface IReplicationBatchEncoder
         }
         var length = BinaryPrimitives.ReadInt32LittleEndian(span[cursor..(cursor + 4)]);
         cursor += 4;
-        if (length < 0 || cursor + length > payload.Length)
+        // Widen to long before summing: length is an attacker-controllable wire
+        // field (up to int.MaxValue), so cursor + length can overflow to a
+        // negative value and slip past this guard, downgrading the precise framing
+        // rejection into a raw slice exception. Mirrors the fix in
+        // OrleansBinaryReplicationBatchEncoder.
+        if (length < 0 || (long)cursor + length > payload.Length)
         {
             throw new ArgumentException(
                 $"Framing payload is truncated at the body for {fieldName}; declared "
